langgraph/pyproject.toml
[project]
name = "agentic-langgraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.0.15",
    "grandalf>=0.8",
    "ipykernel>=6.29.5",
    "langchain>=0.3.25",
    "langchain-community>=0.3.25",
    "langchain-google-genai>=2.1.5",
    "langchain-groq>=0.3.2",
    "langchain-nomic>=0.1.4",
    "langchain-ollama>=0.3.3",
    "langchain-tavily>=0.2.3",
    "langgraph>=0.4.8",
    "langsmith>=0.3.45",
    "python-dotenv>=1.1.0",
]
